Create vpn module and move stuff from lib.rs there
