+++ nym-vpn-core/crates/nym-vpn-cli/src/main.rs
@@ -9,7 +9,8 @@ use nym_vpn_lib::{
     gateway_directory::{Config as GatewayConfig, EntryPoint, ExitPoint},
     nym_bin_common::bin_info,
     nym_config::defaults::{setup_env, var_names},
-    GenericNymVpnConfig, IpPair, MixnetClientConfig, NodeIdentity, NymVpn, Recipient, SpecificVpn,
+    vpn::{GenericNymVpnConfig, MixnetClientConfig, NymVpn, SpecificVpn},
+    IpPair, NodeIdentity, Recipient,
 };
 use time::OffsetDateTime;
 use tracing::{debug, error, info};
