+++ nym-vpn-core/crates/nym-vpnd/src/command_interface/status_broadcaster.rs
@@ -3,7 +3,7 @@
 
 use futures::StreamExt;
 use nym_bandwidth_controller_pre_ecash::BandwidthStatusMessage;
-use nym_vpn_lib::{connection_monitor::ConnectionMonitorStatus, NymVpnStatusMessage};
+use nym_vpn_lib::{connection_monitor::ConnectionMonitorStatus, vpn::NymVpnStatusMessage};
 use nym_vpn_proto::{connection_status_update::StatusType, ConnectionStatusUpdate};
 use tracing::debug;
