+++ nym-vpn-core/nym-vpn-lib/src/mixnet/connect.rs
@@ -8,7 +8,7 @@ use tracing::{debug, info};
 
 use crate::{
     error::{Error, MixnetError, Result},
-    MixnetClientConfig,
+    vpn::MixnetClientConfig,
 };
 
 use super::SharedMixnetClient;
