+++ nym-vpn-core/nym-vpn-lib/src/platform/status_listener.rs
@@ -3,7 +3,7 @@
 
 use crate::platform::{uniffi_set_listener_status, RUNTIME};
 use crate::uniffi_custom_impls::{StatusEvent, TunStatus};
-use crate::NymVpnStatusMessage;
+use crate::vpn::NymVpnStatusMessage;
 use nym_bandwidth_controller_pre_ecash::BandwidthStatusMessage;
 use nym_connection_monitor::ConnectionMonitorStatus;
 use nym_task::manager::{SentStatus, TaskStatus};
